+++ src/com/android/settings/ApnSettings.java
@@ -96,7 +96,7 @@ public class ApnSettings extends SettingsPreferenceFragment implements
     private HandlerThread mRestoreDefaultApnThread;
 
     private UserManager mUm;
-    private long mSubId;
+    private int mSubId;
 
     private String mSelectedKey;
 
@@ -167,7 +167,7 @@ public class ApnSettings extends SettingsPreferenceFragment implements
 
         addPreferencesFromResource(R.xml.apn_settings);
         getListView().setItemsCanFocus(true);
-        mSubId = getActivity().getIntent().getLongExtra(PhoneConstants.SUBSCRIPTION_KEY,
+        mSubId = getActivity().getIntent().getIntExtra(PhoneConstants.SUBSCRIPTION_KEY,
                 SubscriptionManager.getDefaultDataSubId());
         Log.d(TAG, "onCreate received subId :" + mSubId);
         mMobileStateFilter = new IntentFilter(
